requirements.txt
streamlit==1.37.0
openai==1.37.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
crewai==0.51.1